Apply NekoBoxForAndroid.patch + fix build scripts

+++ libcore/build.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
-source ./env_java.sh || true
+[ -f ./env_java.sh ] && source ./env_java.sh
 source ../buildScript/init/env_ndk.sh
 
 BUILD=".build"
@@ -14,8 +14,14 @@ if [ -z "$GOPATH" ]; then
   GOPATH=$(go env GOPATH)
 fi
 
-export GOBIND=gobind-matsuri
-"$GOPATH"/bin/gomobile-matsuri bind -v -androidapi 21 -cache "$(realpath $BUILD)" -trimpath -ldflags='-s -w' -tags='with_conntrack,with_gvisor,with_quic,with_wireguard,with_utls,with_clash_api' . || exit 1
+# Verify gVisor workaround.
+if ! go run -tags=with_gvisor ../../sing-box/cmd/internal/gvisor_workaround_verify/main.go; then
+  echo "ERROR: gVisor workaround verification failed"
+  exit 1
+fi
+
+export GOBIND=gobind
+gomobile bind -v -androidapi 23 -trimpath -ldflags='-s -w' -tags='with_conntrack,with_gvisor,with_quic,with_wireguard,with_awg,with_utls,with_clash_api' . || exit 1
 rm -r libcore-sources.jar
 
 proj=../app/libs
